xsl: CMDI2HTML: fix tab component, broken by e9e7105
Move top level apply-templates back inside <nav>.

I think I moved this out in e9e7105 because semantically, the actual
tab content shouldn't be inside the <nav> element. But this breaks the
JQuery tab component. I'd like to switch to pure CSS tabs, but that's
a project for another day.