CMDI2HTML: a first draft of CSS for nav links
Keeping this in a local <style> tag for now. Will move to a separate
CSS sheet once complete.

The approach here is inspired by the one at
https://www.cssscript.com/creating-simple-tabbed-interface-with-pure-css/

using :target to switch the visibility of the tab contents, but it
still keeps the markup for the section headings separate from the nav
links (i.e., the tab buttons). The problem with this is that clicking
the links jumps to the associated target, thus hiding the nav links.
This is not a problem with the solution linked above, but a
disadvantage of that approach is that there is no separate "table of
contents"-like markup; thus users of browsers without CSS support,
search engines, etc. don't get a set of nav links at the top of the
page.

So I need to think about this a bit more. I want to keep the separate
nav links for non-CSS browsers, but need to find a way to keep the nav
links visible without obscuring the section headings behind them upon jumps.