CMDI2HTML: move general info tab to last position; reduce CSS
I decided to make the compromise of moving General Information to
being the last child <section> in the HTML. This is a compromise: it
means that displaying General Info is somewhat broken for browsers
that do not support CSS. To see General Info in such a browser, the
user will need to click the nav link to get there or scroll to the
bottom. But I do not expect that we will have many such users. I also
believe this should have almost no impact on search engine rankings.

The benefits are that, for users whose browser *does* support CSS, we
get a pure CSS implementation of the nav controls that (a) works with
the back button and (b) doesn't require loading any javascript. It
should thus be very sustainable.

The compromise is necessary because, at least in today's CSS, this is the
only way to both display the General Info tab by default, and stop
displaying it when the user clicks the nav link to select another tab.
The reason is basically that the CSS general sibling selector 'A ~ B'
can only select sibling elements B which *follow* A in the HTML. Thus
the only way to say "display this section by default, but don't
display it when a sibling section is :target" in CSS is to ensure that
the default section appears last in the HTML.

If this turns out to be the wrong compromise, a different option would
be to use z-index to display the :target section on top of the default
section. But this has its own problems (requires absolute positioning,
requires all sections to be at least as large as the default one, etc.).

More info:
https://stackoverflow.com/questions/6906724/is-it-possible-to-have-tabs-without-javascript

A few other cleanups:
  - remove experiment to make tab buttons stick to top of page; will return
    to this later
  - change nav#tabs to nav#toc
  - remove relative positioning for sections