CMDI2HTML: draft flexbox stuff for nav buttons
make the ul in nav#toc a flexbox, distribute space evenly around the
li, add some padding and border to the li so they make nice buttons.

This is just a draft; final styling to be considered in more detail
later, taking responsive design into account, etc.  Might also want to
move to a vertical layout for the nav controls. We might even want
to keep using a CSS framework like Bootstrap or similar. But for now
this restores the general idea that the nav buttons are "tabs" to jump
to the relevant section.