xsl: ExperimentContext: remove old code and refactor new approach
In particular:
  - unused templates have been removed
  - I decided to just follow the structure of the XML for each experiment
  - I'm now using apply-templates with mode=... instead of
  call-template

On the last point: instead of
<xsl:call-template name="FooAsDetails" />
<xsl:template name="FooAsDetails">...

we now have
<xsl:apply-templates select="Foo" mode="details">
<xsl:template match="Foo" mode="details">

which is basically just as explicit, but also changes the context
node, making it easier to make templates more modular and avoid
deeply-nested xpath expressions. I'm just using mode names that are
descriptions of the expected HTML output, e.g. "details".