xsl: Publications: refactor to bibliography-like list
drop old AsTable template

Instead, create an ordered list of each publication, more or less as
it would appear in a bibliography. Unfortunately, we don't record year
or journal here, so we have to rely on the link.

Authors get formatted as "first name last name" for now, as in the old
template, with a link to their ORCID or similar if available.