xsl: ResourceProxyList: use FileSize if provided
Bagman outputs a FileSize element in bytes on ResourceProxyInfo nodes,
rather than a SizeInfo node with TotalSize, etc.
This change makes the stylesheet display the SizeInfo if present,
and also fixes an outstanding TODO about using the SizeUnit, and falls
back to FileSize if it's present but SizeInfo is not.